StudentInfo, SchoolAssignment, Classroom 구조 변경 (현재 학적 추가, 학년,반,학번 현재 학적으로 조회, Classroom이 학생을 물지 않고 쿼리문을 통해 조회하도록 구현)

+++ src/main/java/com/example/schoolmate/cheol/controller/StudentController.java
@@ -43,7 +43,7 @@ public class StudentController {
 
     // 학번으로 학생 상세 조회
     @GetMapping("/details/number/{studentNumber}")
-    public String getStudentDetailsByNumber(@PathVariable Long studentNumber, Model model) {
+    public String getStudentDetailsByNumber(@PathVariable Integer studentNumber, Model model) {
         StudentResponseDTO student = studentService.getStudentByStudentNumber(studentNumber);
         model.addAttribute("student", student);
         return "student/student-details";
